requirements.txt
# bert_score==0.3.13
fastapi==0.115.12
groq==0.26.0
httpx==0.28.1
llama_index==0.12.39
loguru==0.7.3
openai==1.82.1
Pillow==11.2.1
pinecone==7.0.2
pydantic==2.11.5
python-dotenv==1.1.0
SQLAlchemy==2.0.41
swarms==7.8.1
uvicorn==0.34.3
llama-index-embeddings-huggingface==0.5.4
openai-agents==0.0.16
redis==6.2.0
celery==5.5.3
pymysql
PyJWT==2.10.1
cryptography
